+++ src/FluxFlow.Composition/ApplicationModel/ResourceDefinition.cs
@@ -1,0 +1,8 @@
+namespace FluxFlow.Composition.Model;
+
+public abstract class ResourceDefinition
+{
+    private protected ResourceDefinition()
+    {
+    }
+}